:tada: feat: default citizen license
