+++ packages/schemas/src/admin/values/import.ts
@@ -4,6 +4,7 @@ const LICENSE_TYPE_REGEX = /LICENSE|REGISTRATION_STATUS/;
 export const BASE_VALUE_SCHEMA = z.object({
   value: z.string().min(1).max(255),
   licenseType: z.string().max(255).regex(LICENSE_TYPE_REGEX).nullable().optional(),
+  isDefault: z.boolean().nullable().optional(),
 });
 export const BASE_ARR = z.array(BASE_VALUE_SCHEMA).min(1);
